Remove QueryEditorResult component
